+++ OnidirectionalShadowMap/Source/DirectionalLight.h
@@ -8,7 +8,7 @@ public:
 	DirectionalLight();
 
 	DirectionalLight(
-		GLfloat shadowWidth, GLfloat shadowHeight,
+		GLuint shadowWidth, GLuint shadowHeight,
 		GLfloat red, GLfloat green, GLfloat blue,
 		GLfloat intensity, GLfloat dIntencity,
 		GLfloat xDir, GLfloat yDir, GLfloat zDir);
